more blank-canvas api changes

+++ src/Diagrams/Backend/Canvas.hs
@@ -46,22 +46,22 @@
 -- type of option records and rendering results specific to any 
 -- particular backend. For @b ~ Canvas@ and @v ~ R2@, we have
 --
--- > data Options Canvas V2 Float = CanvaseOptions
+-- > data Options Canvas V2 Double = CanvaseOptions
 -- >  { _size :: SizeSpec2D -- ^^ The requested size
 -- >  }
 --
 -- @
--- data family Render Canvas V2 Float = C (RenderM ())
+-- data family Render Canvas V2 Double = C (RenderM ())
 -- @
 --
 -- @
--- type family Result Canvas V2 Float = Canvas ()
+-- type family Result Canvas V2 Double = Canvas ()
 -- @
 --
 -- So the type of 'renderDia' resolves to
 --
 -- @
--- renderDia :: Canvas -> Options Canvas V2 Float -> QDiagram Canvas V2 Float m ->
+-- renderDia :: Canvas -> Options Canvas V2 Double -> QDiagram Canvas V2 Double m ->
 -- Canvas()
 -- @
 --
@@ -116,8 +116,8 @@ data Canvas = Canvas
 
 type B = Canvas
 
-data CanvasState = CanvasState { _accumStyle :: Style V2 Float
-                               , _csPos :: (Float, Float) }
+data CanvasState = CanvasState { _accumStyle :: Style V2 Double
+                               , _csPos :: (Double, Double) }
 
 makeLenses ''CanvasState
 
@@ -133,19 +133,19 @@ liftC = lift
 runRenderM :: RenderM a -> BC.Canvas a
 runRenderM = flip SS.evalStateStackT def
 
-instance Monoid (Render Canvas V2 Float) where
+instance Monoid (Render Canvas V2 Double) where
   mempty  = C $ return ()
   (C c1) `mappend` (C c2) = C (c1 >> c2)
 
-instance Backend Canvas V2 Float where
-  data Render  Canvas V2 Float = C (RenderM ())
-  type Result  Canvas V2 Float = BC.Canvas ()
-  data Options Canvas V2 Float = CanvasOptions
-          { _canvasSize   :: SizeSpec2D Float   -- ^ the requested size
+instance Backend Canvas V2 Double where
+  data Render  Canvas V2 Double = C (RenderM ())
+  type Result  Canvas V2 Double = BC.Canvas ()
+  data Options Canvas V2 Double = CanvasOptions
+          { _canvasSize   :: SizeSpec2D Double   -- ^ the requested size
           }
 
-  renderRTree :: Canvas -> Options Canvas V2 Float -> RTree Canvas V2 Float Annotation 
-                        -> Result Canvas V2 Float
+  renderRTree :: Canvas -> Options Canvas V2 Double -> RTree Canvas V2 Double Annotation 
+                        -> Result Canvas V2 Double
   renderRTree _ _ rt = evalState canvasOutput initialCanvasRenderState
     where
       canvasOutput :: State CanvasRenderState (BC.Canvas ())
@@ -155,12 +155,12 @@ instance Backend Canvas V2 Float where
 
   adjustDia c opts d = adjustDia2D size c opts (d # reflectY)
 
-runC :: Render Canvas V2 Float -> RenderM ()
+runC :: Render Canvas V2 Double -> RenderM ()
 runC (C r) = r
 
-toRender :: RTree Canvas V2 Float Annotation -> Render Canvas V2 Float
+toRender :: RTree Canvas V2 Double Annotation -> Render Canvas V2 Double
 toRender = fromRTree
-  . Node (RStyle (mempty # recommendFillColor (transparent :: AlphaColour Float)))
+  . Node (RStyle (mempty # recommendFillColor (transparent :: AlphaColour Double)))
   . (:[])
   . splitTextureFills
     where
@@ -178,16 +178,16 @@ data CanvasRenderState = CanvasRenderState
 initialCanvasRenderState :: CanvasRenderState
 initialCanvasRenderState = CanvasRenderState
 
-getSize :: Options Canvas V2 Float -> SizeSpec2D Float
+getSize :: Options Canvas V2 Double -> SizeSpec2D Double
 getSize (CanvasOptions {_canvasSize = s}) = s
 
-setSize :: Options Canvas V2 Float -> (SizeSpec2D Float) -> Options Canvas V2 Float
+setSize :: Options Canvas V2 Double -> (SizeSpec2D Double) -> Options Canvas V2 Double
 setSize o s = o {_canvasSize = s}
 
-size :: Lens' (Options Canvas V2 Float)(SizeSpec2D Float)
+size :: Lens' (Options Canvas V2 Double)(SizeSpec2D Double)
 size = lens getSize setSize
 
-move :: (Float, Float) -> RenderM ()
+move :: (Double, Double) -> RenderM ()
 move p = do csPos .= p
 
 save :: RenderM ()
@@ -202,19 +202,19 @@ newPath = liftC $ BC.beginPath ()
 closePath :: RenderM ()
 closePath = liftC $ BC.closePath ()
 
-moveTo :: Float -> Float -> RenderM ()
+moveTo :: Double -> Double -> RenderM ()
 moveTo x y = do
   liftC $ BC.moveTo (x, y)
   move (x, y)
 
-relLineTo :: Float -> Float -> RenderM ()
+relLineTo :: Double -> Double -> RenderM ()
 relLineTo x y = do
   p <- use csPos
   let p' = p + (x, y)
   liftC $ BC.lineTo p'
   move p'
 
-relCurveTo :: Float -> Float -> Float -> Float -> Float -> Float -> RenderM ()
+relCurveTo :: Double -> Double -> Double -> Double -> Double -> Double -> RenderM ()
 relCurveTo ax ay bx by cx cy = do
   p <- use csPos
   let [(ax',ay'),(bx',by'),(cx',cy')] = map (p +) [(ax,ay),(bx,by),(cx,cy)]
@@ -239,7 +239,7 @@ stroke = do
   -- be 'Nothing'. 0.5 is choose since it is the lower bound of the
   -- default.
   w <- fromMaybe 0.5 <$> getStyleAttrib 
-                        (fromOutput . getLineWidth :: LineWidth Float -> Float)
+                        (fromOutput . getLineWidth :: LineWidth Double -> Double)
   when (w > 0) (liftC $ BC.stroke ())
 
 fill :: RenderM ()
@@ -248,12 +248,12 @@ fill = liftC $ BC.fill ()
 clip :: RenderM ()
 clip = liftC $ BC.clip ()
 
-byteRange :: Float -> Word8
+byteRange :: Double -> Word8
 byteRange d = floor (d * 255)
 
 data TextureUse = Fill | Strk
 
-texture :: TextureUse -> Texture Float -> Float -> RenderM()
+texture :: TextureUse -> Texture Double -> Double -> RenderM()
 texture u (SC (SomeColor c))  o = case u of
     Fill -> liftC . S.fillStyle   $ s
     Strk -> liftC . S.strokeStyle $ s
@@ -283,7 +283,7 @@ texture u (RG g) _ = liftC $ do
     stops = map (\st -> ( st^.stopFraction , showColorJS (st^.stopColor) 1)) (g^.rGradStops)
     s = avgScale $ g^.rGradTrans
 
-showColorJS :: (Color c) => c -> Float  -> T.Text
+showColorJS :: (Color c) => c -> Double  -> T.Text
 showColorJS c o = T.concat
     [ "rgba("
     , s (realToFrac r), ","
@@ -292,11 +292,11 @@ showColorJS c o = T.concat
     , T.pack (show $ (realToFrac a) * o)
     , ")"
     ]
-  where s :: Float -> T.Text
+  where s :: Double -> T.Text
         s = T.pack . show . byteRange
         (r,g,b,a) = colorToSRGBA . toAlphaColour $  c
 
-canvasTransform :: T2 Float -> RenderM ()
+canvasTransform :: T2 Double -> RenderM ()
 canvasTransform tr = liftC $ BC.transform vs
     where 
       [[ax, ay], [bx, by], [tx, ty]] = matrixHomRep tr
@@ -304,10 +304,10 @@ canvasTransform tr = liftC $ BC.transform vs
            ,realToFrac bx,realToFrac by
            ,realToFrac tx,realToFrac ty)
 
-strokeTexture :: Texture Float -> Float  -> RenderM ()
+strokeTexture :: Texture Double -> Double  -> RenderM ()
 strokeTexture = texture Strk
 
-fillTexture :: Texture Float -> Float  -> RenderM ()
+fillTexture :: Texture Double -> Double  -> RenderM ()
 fillTexture = texture Fill
 
 fromLineCap :: LineCap -> BC.LineEndCap
@@ -320,7 +320,7 @@ fromLineJoin LineJoinRound = BC.RoundCorner
 fromLineJoin LineJoinBevel = BC.BevelCorner
 fromLineJoin _             = BC.MiterCorner
 
-showFontJS :: FontWeight -> FontSlant -> Float -> String -> T.Text
+showFontJS :: FontWeight -> FontSlant -> Double -> String -> T.Text
 showFontJS wgt slant sz fnt = T.concat [a, " ", b, " ", c, " ", d]
   where
     a = case wgt of
@@ -333,10 +333,10 @@ showFontJS wgt slant sz fnt = T.concat [a, " ", b, " ", c, " ", d]
     c = T.concat [T.pack $ show sz, "pt"]
     d = T.pack fnt
 
-renderC :: (Renderable a Canvas, V a ~ V2, N a ~ Float) => a -> RenderM ()
+renderC :: (Renderable a Canvas, V a ~ V2, N a ~ Double) => a -> RenderM ()
 renderC a = case (render Canvas a) of C r -> r
 
-canvasStyle :: Style v Float  -> RenderM ()
+canvasStyle :: Style v Double  -> RenderM ()
 canvasStyle s = sequence_
               . catMaybes $ [ handle clip' 
                             , handle lWidth
@@ -346,18 +346,18 @@ canvasStyle s = sequence_
   where handle :: (AttributeClass a) => (a -> RenderM ()) -> Maybe (RenderM ())
         handle f = f `fmap` getAttr s
         clip'    = mapM_ (\p -> canvasPath p >> clip) . op Clip
-        lWidth   = liftC . BC.lineWidth . (fromOutput . getLineWidth :: LineWidth Float -> Float)
+        lWidth   = liftC . BC.lineWidth . (fromOutput . getLineWidth :: LineWidth Double -> Double)
         lCap     = liftC . BC.lineCap . fromLineCap . getLineCap
         lJoin    = liftC .  BC.lineJoin . fromLineJoin . getLineJoin
 
-instance Renderable (Segment Closed V2 Float) Canvas where
+instance Renderable (Segment Closed V2 Double) Canvas where
   render _ (Linear (OffsetClosed (V2 x y))) = C $ relLineTo x y
   render _ (Cubic (V2 x1 y1)
                   (V2 x2 y2)
                   (OffsetClosed (V2 x3 y3)))
     = C $ relCurveTo x1 y1 x2 y2 x3 y3
 
-instance Renderable (Trail V2 Float) Canvas where
+instance Renderable (Trail V2 Double) Canvas where
   render _ = withTrail renderLine renderLoop
     where
       renderLine ln = C $ do
@@ -368,7 +368,7 @@ instance Renderable (Trail V2 Float) Canvas where
           _ -> mapM_ renderC (lineSegments . cutLoop $ lp)
         closePath
 
-instance Renderable (Path V2 Float) Canvas where
+instance Renderable (Path V2 Double) Canvas where
   render _ p = C $ do
     canvasPath p
     f <- getStyleAttrib getFillTexture
@@ -376,12 +376,12 @@ instance Renderable (Path V2 Float) Canvas where
     o <- fromMaybe 1 <$> getStyleAttrib getOpacity
     save
     when (isJust f) (fillTexture (fromJust f) (realToFrac o) >> fill)
-    strokeTexture (fromMaybe (SC (SomeColor (black :: Colour Float))) s) (realToFrac o)
+    strokeTexture (fromMaybe (SC (SomeColor (black :: Colour Double))) s) (realToFrac o)
     stroke
     restore
 
 -- Add a path to the Canvas context, without stroking or filling it.
-canvasPath :: Path V2 Float -> RenderM ()
+canvasPath :: Path V2 Double -> RenderM ()
 canvasPath (Path trs) = do
     newPath
     F.mapM_ renderTrail trs
@@ -390,14 +390,14 @@ canvasPath (Path trs) = do
       uncurry moveTo p
       renderC tr
 
-instance Renderable (Text Float) Canvas where
+instance Renderable (Text Double) Canvas where
   render _ (Text tt tn al str) = C $ do
-    isLocal <- fromMaybe True <$> getStyleAttrib (getFontSizeIsLocal :: FontSize Float -> Bool)
+    isLocal <- fromMaybe True <$> getStyleAttrib (getFontSizeIsLocal :: FontSize Double -> Bool)
     tf      <- fromMaybe "Calibri" <$> getStyleAttrib getFont
     sz      <- fromMaybe 12 <$> getStyleAttrib (fromOutput . getFontSize)
     slant   <- fromMaybe FontSlantNormal <$> getStyleAttrib getFontSlant
     fw      <- fromMaybe FontWeightNormal <$> getStyleAttrib getFontWeight
-    tx      <- fromMaybe (SC (SomeColor (black :: Colour Float)))
+    tx      <- fromMaybe (SC (SomeColor (black :: Colour Double)))
                <$> getStyleAttrib getFillTexture
     o       <- fromMaybe 1 <$> getStyleAttrib getOpacity
     let fSize = if isLocal
@@ -425,7 +425,7 @@ instance Renderable (Text Float) Canvas where
     liftC $ BC.fillText (T.pack str, 0, 0)
     restore
 
-instance Renderable (DImage Float External) Canvas where
+instance Renderable (DImage Double External) Canvas where
   render _ (DImage path w h tr) = C $ do
     let ImageRef file = path
     save
@@ -434,7 +434,7 @@ instance Renderable (DImage Float External) Canvas where
     liftC $ BC.drawImage (img, [fromIntegral (-w) / 2, fromIntegral (-h) / 2, fromIntegral w, fromIntegral h])
     restore
 
-renderCanvas :: Int -> SizeSpec2D Float -> Diagram Canvas V2 Float -> IO ()
+renderCanvas :: Int -> SizeSpec2D Double -> Diagram Canvas V2 Double -> IO ()
 renderCanvas port sizeSpec d = BC.blankCanvas (fromIntegral port) . flip BC.send $ img
     where
       img = renderDia Canvas (CanvasOptions sizeSpec) d
